finished mytrips page
